refactor(team-metrics)!: radically change how metrics are stored
We used to store a row per day with an array for each quarter of an
hour, but we now store each quarter on a separated row. This will ease
the timezone support, and makes the stored/returned data more reliable
in case the API was not able to compute (and store) some metrics at a
given time.