refactor(meilisearch): make client injectable & refactor indexer
Add nestjs-meilisearch to have the client injectable in the DI context.

Rewrote the indexer (previously "MeiliSearchGlobal", now
"MeiliSearchIndexerService"), so it is more optimized, with a better
separation of concerns, and easier to read.
(namely: moved interfaces out of the file, move helper function as
private methods so they can benefit from the instance's properties,
made most of the properties private + readonly, remove unused/dead
code...)